[refactor/#221] 서비스 구조 수정
