--- src/main/java/com/clokey/server/domain/member/application/AppleAuthService.java
@@ -1,10 +1,0 @@
-package com.clokey.server.domain.member.application;
-
-import com.clokey.server.domain.member.domain.entity.Member;
-import com.clokey.server.domain.member.dto.AuthDTO;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface AppleAuthService {
-
-    AuthDTO.TokenResponse login(String code, String deviceToken);
-}
