--- src/main/java/com/clokey/server/domain/member/application/IDQueryService.java
@@ -1,6 +1,0 @@
-package com.clokey.server.domain.member.application;
-
-
-public interface IDQueryService {
-}
-
