--- src/main/java/com/clokey/server/domain/member/application/IDQueryServiceImpl.java
@@ -1,12 +1,0 @@
-package com.clokey.server.domain.member.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class IDQueryServiceImpl implements IDQueryService {
-
-
-}
-
