--- src/main/java/com/clokey/server/domain/member/application/ProfileCommandServiceImpl.java
@@ -1,57 +1,0 @@
-package com.clokey.server.domain.member.application;
-
-import com.clokey.server.domain.member.converter.ProfileConverter;
-import com.clokey.server.domain.member.domain.entity.Member;
-import com.clokey.server.domain.model.entity.enums.RegisterStatus;
-import com.clokey.server.domain.member.dto.MemberDTO;
-import com.clokey.server.domain.search.application.SearchRepositoryService;
-import com.clokey.server.domain.search.exception.SearchException;
-import com.clokey.server.global.error.code.status.ErrorStatus;
-import com.clokey.server.global.infra.s3.S3ImageService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-@Service
-@RequiredArgsConstructor
-public class ProfileCommandServiceImpl implements ProfileCommandService {
-
-    private final MemberRepositoryService memberRepositoryService;
-    private final S3ImageService s3ImageService; // ✅ S3 업로드 서비스 추가
-    private final SearchRepositoryService searchRepositoryService;
-
-    @Override
-    @Transactional
-    public MemberDTO.ProfileRP updateProfile(Long userId, MemberDTO.ProfileRQ request,
-                                             MultipartFile profileImage, MultipartFile profileBackImage) {
-        // 사용자 확인
-        Member member = memberRepositoryService.findMemberById(userId);
-
-        // ✅ S3 업로드 후 URL 저장
-        String profileImageUrl = (profileImage != null && !profileImage.isEmpty()) ? s3ImageService.upload(profileImage) : member.getProfileImageUrl();
-        String profileBackImageUrl = (profileBackImage != null && !profileBackImage.isEmpty()) ? s3ImageService.upload(profileBackImage) : member.getProfileBackImageUrl();
-
-        member.profileUpdate(request, profileImageUrl, profileBackImageUrl);
-
-        if (member.getRegisterStatus() != RegisterStatus.REGISTERED) {
-            // 약관 동의가 완료되었으므로 회원의 등록 상태를 업데이트
-            member.updateRegisterStatus(RegisterStatus.REGISTERED);
-        }
-
-        // 저장
-        Member updatedMember = memberRepositoryService.saveMember(member);
-
-        // ES 동기화
-        try {
-            searchRepositoryService.updateMemberDataToElasticsearch(updatedMember);
-        } catch (IOException e) {
-            throw new SearchException(ErrorStatus.ELASTIC_SEARCH_SYNC_FAULT);
-        }
-
-        // 응답 생성
-        return ProfileConverter.toProfileRPDTO(updatedMember);
-    }
-}
